DB code restructure and module renaming
