+++ lib/core/cardano-wallet-core.cabal
@@ -195,17 +195,22 @@ library
       Cardano.Wallet.CoinSelection.Internal.Context
       Cardano.Wallet.Compat
       Cardano.Wallet.DB
-      Cardano.Wallet.DB.Checkpoints
       Cardano.Wallet.DB.MVar
       Cardano.Wallet.DB.Model
-      Cardano.Wallet.DB.Sqlite
-      Cardano.Wallet.DB.Sqlite.AddressBook
-      Cardano.Wallet.DB.Sqlite.CheckpointsOld
+      Cardano.Wallet.DB.Layer
+      Cardano.Wallet.DB.Unstored
+      Cardano.Wallet.DB.Store.Checkpoints.AddressBook
+      Cardano.Wallet.DB.Store.Checkpoints.Model
+      Cardano.Wallet.DB.Store.Checkpoints.Persist
+      Cardano.Wallet.DB.Store.TxHistory.Types
+      Cardano.Wallet.DB.Store.TxHistory.Persist
+      Cardano.Wallet.DB.Store.TxHistory.Update
+      Cardano.Wallet.DB.Store.Wallets
       Cardano.Wallet.DB.Sqlite.Migration
-      Cardano.Wallet.DB.Sqlite.Stores
-      Cardano.Wallet.DB.Sqlite.TH
+      Cardano.Wallet.DB.Sqlite.Schema
       Cardano.Wallet.DB.Sqlite.Types
-      Cardano.Wallet.DB.WalletState
+      Cardano.Wallet.DB.Sqlite.WrapSTM
+      Cardano.Wallet.DB.State
       Cardano.Wallet.Logging
       Cardano.Wallet.Network
       Cardano.Wallet.Network.Light
